feat: categories test
