+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "start": "tsc && node dist/src/server.js",
     "dev": "nodemon src/server.ts",
-    "test": "dotenv -e .env.test prisma migrate dev && prisma db seed && cross-env && NODE_OPTIONS=--experimental-vm-modules dotenv -e .env.test jest -- -i"
+    "test": "dotenv -e .env.test prisma migrate reset && prisma db seed && cross-env && NODE_OPTIONS=--experimental-vm-modules dotenv -e .env.test jest -- -i"
   },
   "prisma": {
     "seed": "ts-node prisma/seed.ts"
